requirements.txt
uvicorn
fastapi
requests
Pillow
aiofiles